database conversations hoplite configuration

+++ composeApp/src/commonMain/kotlin/Screen/ConvScreen.kt
@@ -49,7 +49,7 @@ val ConvScreen by navDestination<Unit> {
 }
 
 @Composable
-fun ConvCard(id: Int, name: String, navController: NavController) {
+fun ConvCard(id: Long, name: String, navController: NavController) {
     Row (Modifier.padding(20.dp).clickable { navController.replace(ChatScreen, ChatParams(id, name)) }){
         Image(
             painter = painterResource(Res.drawable.gnomey),
